Renamed test packages to be properly picked by sonar
