+++ tests/org.preesm.algorithm.tests/src/org/preesm/algorithm/synthesis/test/ScheduleIteratorTest.java
@@ -34,7 +34,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
-package org.preesm.algorithm.synthesis;
+package org.preesm.algorithm.synthesis.test;
 
 import static org.junit.Assert.assertEquals;
